DSTORE-359 Close HdfsTransactionLog streams as soon as the log will not be written-to anymore.
Previously, all HDFS tlog files were kept open for writes, which consumes two threads on the HDFS DataNode
for the DataXceiver and PacketResponder. This is wasteful as most tlogs will not be written to, and worse
is that Solr *must not* write to a completed tlog. Now we close once they are done but keep them available
for reads e.g. recovery, real-time gets, peer sync.