CAPI-571 Fix SolrJ Convert Child Docs Correctly
The current solrj Document DTO to SolrInputDocument binder
treat all child docs as anonymous rather than prefix them with
the given path. This will result in the document being indexed
without the `_nest_path_` field populated which is required
for the ChildDocumentTransformer (`fl=[child]`) to work.

In order to fix this if the Document DTO contain any fields
which are also a DTO or a List/Array of DTOs, they should be
converted as SolrInputDocuments on its own and nested as sub
document(s) within the parent document.